requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==5.0
djangorestframework==3.14.0
Faker==22.5.1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pip==23.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.27.0